+++ at_secondary/at_persistence_secondary_server/lib/src/notification/at_notification_keystore.dart
@@ -9,7 +9,7 @@ import 'package:hive/hive.dart';
 /// Class to initialize, put and get entries into [AtNotificationKeystore]
 class AtNotificationKeystore
     with HiveBase<AtNotification?>
-    implements SecondaryKeyStore {
+    implements SecondaryKeyStore, AtLogType {
   static final AtNotificationKeystore _singleton =
       AtNotificationKeystore._internal();
 
@@ -19,6 +19,8 @@ class AtNotificationKeystore
   factory AtNotificationKeystore.getInstance() {
     return _singleton;
   }
+
+  final _logger = AtSignLogger('AtNotificationKeystore');
 
   bool _register = false;
 
@@ -81,20 +83,55 @@ class AtNotificationKeystore
       bool? isBinary,
       bool? isEncrypted,
       String? dataSignature}) async {
-    // TODO: implement deleteExpiredKeys
     throw UnimplementedError();
   }
 
   @override
   Future<bool> deleteExpiredKeys() async {
-    // TODO: implement deleteExpiredKeys
-    return Future.value(false);
+    var result = true;
+    try {
+      var expiredKeys = await getExpiredKeys();
+      if (expiredKeys.isNotEmpty) {
+        _logger.finer('expired keys: $expiredKeys');
+        await Future.forEach(expiredKeys, (expiredKey) async {
+          await remove(expiredKey);
+        });
+      } else {
+        _logger.finer('notification key store. No expired notifications');
+      }
+    } on Exception catch (e) {
+      result = false;
+      _logger.severe('Exception in deleteExpired keys: ${e.toString()}');
+      throw DataStoreException(
+          'exception in deleteExpiredKeys: ${e.toString()}');
+    } on HiveError catch (error) {
+      _logger.severe('HiveKeystore get error: $error');
+      throw DataStoreException(error.message);
+    }
+    return result;
   }
 
   @override
   Future<List<dynamic>> getExpiredKeys() async {
-    // TODO: implement getExpiredKeys
-    return <dynamic>[];
+    var expiredKeys = <String>[];
+    try {
+      var keys = _getBox().keys;
+      var expired = [];
+      await Future.forEach(keys, (key) async {
+        var value = await get(key);
+        if (value != null && value.isExpired()) {
+          expired.add(key);
+        }
+      });
+      expired.forEach((key) => expiredKeys.add(Utf7.encode(key)));
+    } on Exception catch (e) {
+      _logger.severe('exception in hive get expired keys:${e.toString()}');
+      throw DataStoreException('exception in getExpiredKeys: ${e.toString()}');
+    } on HiveError catch (error) {
+      _logger.severe('HiveKeystore get error: $error');
+      throw DataStoreException(error.message);
+    }
+    return expiredKeys;
   }
 
   @override
@@ -114,24 +151,6 @@ class AtNotificationKeystore
     }
     encodedKeys?.forEach((key) => keys.add(Utf7.decode(key)));
     return encodedKeys;
-  }
-
-  @override
-  Future getMeta(key) {
-    // TODO: implement getMeta
-    throw UnimplementedError();
-  }
-
-  @override
-  Future putAll(key, value, metadata) {
-    // TODO: implement putAll
-    throw UnimplementedError();
-  }
-
-  @override
-  Future putMeta(key, metadata) {
-    // TODO: implement putMeta
-    throw UnimplementedError();
   }
 
   @override
@@ -157,7 +176,57 @@ class AtNotificationKeystore
 
   @override
   bool isKeyExists(String key) {
-    // TODO: implement isKeyExists
+    return _getBox().keys.contains(key);
+  }
+
+  @override
+  Future<void> delete(expiredKeys) async {
+    try {
+      if (expiredKeys.isNotEmpty) {
+        _logger.finer('expired keys: $expiredKeys');
+        await Future.forEach(expiredKeys, (expiredKey) async {
+          await remove(expiredKey);
+        });
+      } else {
+        _logger.finer('notification key store. No expired notifications');
+      }
+    } on Exception catch (e) {
+      _logger.severe('Exception in deleteExpired keys: ${e.toString()}');
+      throw DataStoreException(
+          'exception in deleteExpiredKeys: ${e.toString()}');
+    } on HiveError catch (error) {
+      _logger.severe('HiveKeystore get error: $error');
+      throw DataStoreException(error.message);
+    }
+  }
+
+  @override
+  int entriesCount() {
+    return _getBox().keys.length;
+  }
+
+  @override
+  Future<List> getExpired(int expiryInDays) async {
+    return await getExpiredKeys();
+  }
+
+  @override
+  Future getMeta(key) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future putAll(key, value, metadata) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future putMeta(key, metadata) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<List> getFirstNEntries(int N) {
     throw UnimplementedError();
   }
 }