+++ at_secondary/at_secondary_server/test/notify_list_verb_test.dart
@@ -11,6 +11,7 @@ import 'package:at_secondary/src/verb/handler/notify_list_verb_handler.dart';
 import 'package:at_server_spec/at_verb_spec.dart';
 import 'package:crypto/crypto.dart';
 import 'package:test/test.dart';
+import 'package:uuid/uuid.dart';
 
 void main() {
   var storageDir = Directory.current.path + '/test/hive';
@@ -132,6 +133,8 @@ void main() {
       expect('@test_user_1', result[0]['from']);
       expect('@bob', result[0]['to']);
       expect('key-3', result[0]['key']);
+      await AtNotificationKeystore.getInstance().remove('122');
+      await AtNotificationKeystore.getInstance().remove('125');
     });
 
     test('A test to verify from and to date', () async {
@@ -221,6 +224,139 @@ void main() {
       expect('@test_user_1', result[1]['from']);
       expect('@bob', result[1]['to']);
       expect('key-2', result[1]['key']);
+      await AtNotificationKeystore.getInstance().remove('121');
+      await AtNotificationKeystore.getInstance().remove('122');
+      await AtNotificationKeystore.getInstance().remove('123');
+    });
+    tearDown(() async => await tearDownFunc());
+  });
+  group('A group of tests on expiry ', () {
+    setUp(() async => keyStoreManager = await setUpFunc(storageDir));
+    test('A test to verify notify list does not return expired entries - 1 expired entry', () async {
+      var notifyListVerbHandler =
+      NotifyListVerbHandler(keyStoreManager.getKeyStore());
+      var notification1 = (AtNotificationBuilder()
+        ..id = '122'
+        ..fromAtSign = '@test_user_1'
+        ..notificationDateTime = DateTime.now()
+        ..toAtSign = '@bob'
+        ..notification = 'key-2'
+        ..type = NotificationType.received
+        ..opType = OperationType.update
+        ..messageType = MessageType.key
+        ..expiresAt = null
+        ..priority = NotificationPriority.low
+        ..notificationStatus = NotificationStatus.queued
+        ..retryCount = 0
+        ..strategy = 'latest'
+        ..notifier = 'persona'
+        ..depth = 3..ttl=100)
+          .build();
+
+      var notification2 = (AtNotificationBuilder()
+        ..id = '125'
+        ..fromAtSign = '@test_user_1'
+        ..notificationDateTime = DateTime.now()
+        ..toAtSign = '@bob'
+        ..notification = 'key-3'
+        ..type = NotificationType.received
+        ..opType = OperationType.update
+        ..messageType = MessageType.key
+        ..expiresAt = null
+        ..priority = NotificationPriority.low
+        ..notificationStatus = NotificationStatus.queued
+        ..retryCount = 0
+        ..strategy = 'latest'
+        ..notifier = 'persona'
+        ..depth = 3)
+          .build();
+
+      await AtNotificationKeystore.getInstance().put('122', notification1);
+      await AtNotificationKeystore.getInstance().put('125', notification2);
+      sleep(Duration(milliseconds: 500));
+      var verb = NotifyList();
+      var command = 'notify:list';
+      var regex = verb.syntax();
+      var verbParams = getVerbParam(regex, command);
+      var inBoundSessionId = '123';
+      var metadata = InboundConnectionMetadata()
+        ..fromAtSign = '@alice'
+        ..isAuthenticated = true;
+      var atConnection = InboundConnectionImpl(null, inBoundSessionId)
+        ..metaData = metadata;
+      var response = Response();
+      await notifyListVerbHandler.processVerb(
+          response, verbParams, atConnection);
+      var result = jsonDecode(response.data!);
+      print(result);
+      expect(result.length, 1);
+      expect(result[0]['id'], '125');
+      await AtNotificationKeystore.getInstance().remove('122');
+      await AtNotificationKeystore.getInstance().remove('125');
+    });
+
+    test('A test to verify notify list expired entries - No expired entry', () async {
+      var notifyListVerbHandler =
+      NotifyListVerbHandler(keyStoreManager.getKeyStore());
+      var notification1 = (AtNotificationBuilder()
+        ..id = '122'
+        ..fromAtSign = '@test_user_1'
+        ..notificationDateTime = DateTime.now()
+        ..toAtSign = '@bob'
+        ..notification = 'key-2'
+        ..type = NotificationType.received
+        ..opType = OperationType.update
+        ..messageType = MessageType.key
+        ..expiresAt = null
+        ..priority = NotificationPriority.low
+        ..notificationStatus = NotificationStatus.queued
+        ..retryCount = 0
+        ..strategy = 'latest'
+        ..notifier = 'persona'
+        ..depth = 3..ttl=60000)
+          .build();
+
+      var notification2 = (AtNotificationBuilder()
+        ..id = '125'
+        ..fromAtSign = '@test_user_1'
+        ..notificationDateTime = DateTime.now()
+        ..toAtSign = '@bob'
+        ..notification = 'key-3'
+        ..type = NotificationType.received
+        ..opType = OperationType.update
+        ..messageType = MessageType.key
+        ..expiresAt = null
+        ..priority = NotificationPriority.low
+        ..notificationStatus = NotificationStatus.queued
+        ..retryCount = 0
+        ..strategy = 'latest'
+        ..notifier = 'persona'
+        ..depth = 3..ttl=70000)
+          .build();
+
+      await AtNotificationKeystore.getInstance().put('122', notification1);
+      await AtNotificationKeystore.getInstance().put('125', notification2);
+      sleep(Duration(milliseconds: 500));
+      var verb = NotifyList();
+      var command = 'notify:list';
+      var regex = verb.syntax();
+      var verbParams = getVerbParam(regex, command);
+      var inBoundSessionId = '123';
+      var metadata = InboundConnectionMetadata()
+        ..fromAtSign = '@alice'
+        ..isAuthenticated = true;
+      var atConnection = InboundConnectionImpl(null, inBoundSessionId)
+        ..metaData = metadata;
+      var response = Response();
+      await notifyListVerbHandler.processVerb(
+          response, verbParams, atConnection);
+      var result = jsonDecode(response.data!);
+      print(result);
+      expect(result.length, 2);
+      expect(result[0]['id'],'122');
+      expect(result[1]['id'], '125');
+      await AtNotificationKeystore.getInstance().remove('122');
+      await AtNotificationKeystore.getInstance().remove('125');
     });
     tearDown(() async => await tearDownFunc());
   });