+++ at_secondary/at_persistence_secondary_server/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.10
+- persistence spec version upgrade
+## 3.0.9
+- Added support for notification expiry based on ttl
 ## 3.0.8
 - at_utils and at_commons version upgrade.
 ## 3.0.7